packages/workers/chucker/pyproject.toml
[tool.poetry]
name = "insighthub-worker-chucker"
version = "0.1.0"
description = "Document chunker worker - splits documents into smaller chunks"
authors = ["Nate Stott <nate.stott@pm.me>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
nltk = "^3.9.0"
insighthub-shared = {path = "../../shared/python/dist/insighthub_shared-0.1.1-py3-none-any.whl"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
mypy = "^1.8.0"
ruff = "^0.2.0"
testcontainers = "^4.13.3"
pytest-cov = "^7.0.0"
black = "^25.11.0"
isort = "^7.0.0"
pytest-mock = "^3.15.1"
types-psycopg2 = "^2.9.21.20251012"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
